(#11324) Created workshop - DBA Essentials Workshops Series
Hello,

Developed a new workshop on DBA Essentials Workshops Series for Oracle Database 21c, WMS ID is #11324.

We got a series of DBA Essentials workshops published on Livelabs. This workshop acts as a master of all these workshops. It introduces the workshops from this series and helps the user run them in a specific sequence.

Please approve the PR. Thank you
~MG