Changed Titles for Lab 2 and 3 in manifest
I ran the Lintchecker on the workshop. Please make these minor changes and create the PR when the content is ready to push to production.
Your lab: 'Lab 2: Setting up the In-Memory Column Store', uses a gerund. Consider using an imperative verb lab title instead of a gerund, for example, 'Start' instead of 'Starting'.
Your lab: 'Lab 3: Querying the In-Memory Column Store', uses a gerund. Consider using an imperative verb lab title instead of a gerund, for example, 'Start' instead of 'Starting'.