Pull request for in-memory-21c (#265)
* New In-Memory 21c Lab

New In-Memory 21c Lab

* Removed image files

Removed image files and update md files with text to replace images.

* Updated misc files

Updated manifest.json files for new labs. Only Desktop was up to date.
Update README.md for new lab description
Fixed setup.md with correct row number for Lineorder table.

* Added proper text for moving to next lab

Added the proper text for moving to the next lab at the end of each lab's markdown file.

* Minor updates to syntax and wording

* Minor text updates

Changed the text around the videos

* Update joins-aggregations.md

Remove invalid comments for old video link and description.

* Fixed typo and added Lab scripts

Fixed a comment typo in joins-aggregations.md and added a script directory with all of the Lab scripts.

* Updated some wording and examples.

Made a final pass on these Labs updating some wording and examples.

* Moved scripts to a files directory under each lab

* Changes due to AIM issues

Changed three scripts to make the AIM High demo more robust and corrected one script name in the md file.

* Fixed Pre-Req section

Fixed the pre-requisites section so the required labs are properly detailed.

* Minor typing corrections

Corrected a couple of words in the Queries lab

* Update estimated lab times

Updated the estimated time for each lab and total lab time in the introduction.

* Added IME lab

This lab shows how In-Memory Expressions work

* Added Join Groups Lab

Added lab to show In-Memory Join Groups and their performance benefit

* Added IM Arithmetic lab

This lab shows how In-Memory Optimized Arithmetic works

* Added External Table Lab

Added lab to show the support for external tables with Database In-Memory

* Add In-Memory Full Text Columns Lab

Add the In-Memory Full Text Columns lab

* Added ADO, Spatial and Hybrid Scans labs

Added ADO, Spatial and Hybrid Scans labs and corrected a couple of the other markdown files

* Add AIM and Parallel Features labs

Add the AIM and Parallel Features labs

* Updated formatting on md files

Updated formatting and fixed IMDS SQL file

* Testing updates

Renamed directories to match lab on disk directories and corrected a couple of issues with parameter settings and unpopulating the lineorder table to make room for other lab objects.

* Update 03_enable_arith.sql

Changed enable to just memory, not spfile

* Last SQL changes based on final review

Last SQL changes to the advanced labs based on a final review

* Updated workshop files

Updated workshop files and added ocw23 folders

* Removed SQL and CSV files

Removed the SQL and CSV files for the labs.

* Removed old in-memory directory

Removed old in-memory directory. No longer used as everything transitioned to the in-memory-21c directory last year.

* Changed Titles for Lab 2 and 3 in manifest

I ran the Lintchecker on the workshop. Please make these minor changes and create the PR when the content is ready to push to production.
Your lab: 'Lab 2: Setting up the In-Memory Column Store', uses a gerund. Consider using an imperative verb lab title instead of a gerund, for example, 'Start' instead of 'Starting'.
Your lab: 'Lab 3: Querying the In-Memory Column Store', uses a gerund. Consider using an imperative verb lab title instead of a gerund, for example, 'Start' instead of 'Starting'.

* Added video links for labs and introductions

Added video links for lab walk throughs and updated some of the lab introduction videos.

* Updated manifest.json for Advanced Lab

There were three directories with the wrong names.

---------

Co-authored-by: Andy Rivenes <andyr.github@gmail.com>
Co-authored-by: Andy Rivenes <andy.rivenes@gmail.com>