LWMS ID 11389: Inithial creation from development folder. Made change… (#293)
* LWMS ID 11389: Inithial creation from development folder. Made changes recommended by Adrienne Howard

LWMS ID 11389: Inithial creation from development folder. Made changes recommended by Adrienne Howard
1.	 Before VCN -do we need to create a compartment?  Generally root should not be used -correct? -done
2.	Lab 1 Task 4  - text says compartment rood -image shows compartment heatwave
3.	Is the PAR URL correct? ???
4.	Create took about 30 minutes   ??
5.	Lab 2 Task 3 -says to keep the shape -but then we change it?  Also the screens shots are out of date so it is a bit confusing. - done
6.	Lab 3, Task 3 Step 4 there is a type-should be 3 not 31  -done
I could not enable the MySql HeatWave Lakehouse Option. -  done
8.	Adding the cluster takes about 30 minutes
9.	Had to disable point in time backups to enable lakehouse
10.	 This is the correct statement for Lab 4 Task 2 step 3 SELECT table_name, table_rows FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_SCHEMA = 'mysql_customer_orders';
Lab 6  – screen shots of the query results would help – if we have time
Delivery-vendor.pq is in the wrong folder -causes ALTER TABLE to fail.
Lab 7 – some screen shots of results would be nice if we have time -but all instructions work.
At the beginning of each lab -it would be nice to re-include the instructions on how to connect to the compute node and then the MySQL Db.  If you get disconnected – it is hard to find the instructions on how to do this

* LWMS ID 11389: Made changes requested by Hope Fisher

Pease update the image alternate text to more accurately match the image. This is the text within ![ ]