(#11767) New workshop from DBA Essentials - install database 23ai
Hello,

We have developed a new workshop - Install Oracle Database 23ai on OCI Compute, WMS ID is #11767.

Oracle Database 23ai (23.6) GA was Oct 15, Tue. Therefore, to compliment the database release, we would like to publish our workshops from the DBA Essentials series for 23ai.
Would be great if you approve this PR please so that we could proceed with publishing to production on priority.
Remaining workshops from the series would follow suit and we intend to publish them all in the next 2 weeks.

Thanks a lot for your help and support.

~MG