LWMS ID: 11389 Updated with @carmenberdant1 requested changes on this pull request.
Lab 1 - task 2, step 2 the original compartment is showing instead of the one in the example. Same for the Lab 1 - Task 5 step 2
Lab 3 - Task 2 - step 6 needs to have all the places where the root and compartment shows, blurred.
Lab 4 - Task 1 step 4 point b) there is a typo for the word command
Lab 4 - Task 2 step 3 - the original compartment is showing